third_party/blink/web_tests/external/wpt/css/css-page/parsing/size-valid-expected.txt

This is a testharness.js-based test.
[FAIL] size: letter
  assert_equals: for rule 12 expected "letter;" but got "letter portrait;"
Harness: the test ran to completion.
